add a collection of embedded youtube videos
Add a jekyll collection for videos.
Add the 2 published videos from the youtube channel.
Add a main videos page with a list of all videos.
Add a video layout which embeds the video, adds a link to youtube, and the video description.
Add a youtube channel social media icon in the footer.